requirements.txt
fastapi==0.100.0
httpx==0.24.1
pydantic==1.10.9
uvicorn==0.22.0
sqlalchemy==2.0.18
passlib==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
python-multipart==0.0.6
email-validator==1.3.1
bcrypt==4.0.1
fastapi-mail==1.3.1
